contrib/hamilton/contrib/user/elijahbenizzy/convert_images_s3/requirements.txt
boto3
pillow
